requirements.txt
# Core ML/AI
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Vector operations
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Graph database
neo4j>=5.14.0
py2neo>=2021.2.3

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# MCP integration
mcp>=0.1.0
playwright>=1.40.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
ruff>=0.1.6
mypy>=1.7.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipykernel>=6.26.0
matplotlib>=3.8.0
